requirements.txt
chainlit
openai
python-dotenv
langchain
langchain-community
pypdf
-e .